feat: lobbyLayout 으로 분리
